requirements.txt
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
tree-sitter-python==0.23.6
python-dotenv
tiktoken==0.8.0
python-scalpel
hypothesis-jsonschema==0.23.1
openai==1.58.1
coverage==7.8.0
python-dotenv==1.0.1
python_graphs
source_parser
javalang